+++ src/lib/schema/runtime.ts
@@ -8,7 +8,7 @@ export interface ExecObject {
   };
 }
 
-export type ExecCommand = string | ExecObject | any[]; // should be ExecCommand[], but that gives a circular type
+export type ExecCommand = string | ExecObject | (string | ExecObject)[];
 
 export type ExecFunction = (...commands: ExecCommand[]) => Promise<{ keepContext: ExecFunction }>;